+++ crates/forge/src/scarb/config.rs
@@ -1,5 +1,6 @@
 use anyhow::{bail, Result};
 use conversions::StarknetConversions;
+use forge_runner::ForkTarget;
 use itertools::Itertools;
 use serde::Deserialize;
 use starknet::core::types::{BlockId, BlockTag};
@@ -17,12 +18,6 @@ pub struct ForgeConfig {
     pub fuzzer_seed: Option<u64>,
 
     pub fork: Vec<ForkTarget>,
-}
-
-#[derive(Debug, PartialEq, Clone)]
-pub struct ForkTarget {
-    pub name: String,
-    pub params: RawForkParams,
 }
 
 /// Represents forge config deserialized from Scarb.toml using basic types like String etc.
